HADOOP-19608. Upgrade to JUnit 5.13.3 + Surefire 3.5.3 (#7785)
JUnit has been upgraded to 5.13.3 so that test suites can
be parameterized at the class level (again).
This requires a matching upgrade to Surefire and some
tuning of surefire options to restore existing behavior.

Dependency Changes:
* junit.jupiter and junit.vintage => 5.13.3
* junit.platform => 1.13.3
* Surefire => 3.5.3.

Changed Surefire Flags: 
* trimStackTrace => false
* surefire.failIfNoSpecifiedTests => false

Build property trimStackTrace
-----------------------------

trimStackTrace is set to false by default

This restores the behavior that any test failure
includes a deep stack trace rather than the first two elements.
In this version of surefire it also seems to print the full stack
of JUnit test execution and the java runtime itself.

This is only an issue for when tests fail.
If the short trace is desired, enable it on the command line:

    mvn test -DtrimStackTrace

Build property surefire.failIfNoSpecifiedTests
----------------------------------------------

surefire.failIfNoSpecifiedTests is set to false by default    

This is required to retain the behavior where the invocation
    mvn verify -Dtest=unknown
will skip the unit tests but still run the integration tests, and a
specific property "it.test" can name the specific test to run:

    mvn verify -Dtest=unknown -Dit.test=ITestConnection

Contributed by Steve Loughran